Make Tumble view control rotate on drag
