+++ PartPreviewWindow/View3D/InteractionLayer.cs
@@ -30,168 +30,17 @@ either expressed or implied, of the FreeBSD Project.
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Threading;
-using System.Threading.Tasks;
 using MatterHackers.Agg;
-using MatterHackers.Agg.Image;
 using MatterHackers.Agg.UI;
 using MatterHackers.MeshVisualizer;
-using MatterHackers.PolygonMesh;
 using MatterHackers.RayTracer;
 using MatterHackers.RayTracer.Traceable;
-using MatterHackers.RenderOpenGl;
 using MatterHackers.RenderOpenGl.OpenGl;
 using MatterHackers.VectorMath;
 using static MatterHackers.MeshVisualizer.MeshViewerWidget;
 
 namespace MatterHackers.MatterControl.PartPreviewWindow
 {
-	public class LightingData
-	{
-		internal float[] ambientLight = { 0.2f, 0.2f, 0.2f, 1.0f };
-
-		internal float[] diffuseLight0 = { 0.7f, 0.7f, 0.7f, 1.0f };
-		internal float[] specularLight0 = { 0.5f, 0.5f, 0.5f, 1.0f };
-		internal float[] lightDirection0 = { -1, -1, 1, 0.0f };
-
-		internal float[] diffuseLight1 = { 0.5f, 0.5f, 0.5f, 1.0f };
-		internal float[] specularLight1 = { 0.3f, 0.3f, 0.3f, 1.0f };
-		internal float[] lightDirection1 = { 1, 1, 1, 0.0f };
-	}
-
-	public class TumbleCubeControl : GuiWidget
-	{
-		LightingData lighting = new LightingData();
-		Mesh cube = PlatonicSolids.CreateCube(3, 3, 3);
-		IPrimitive cubeTraceData;
-		InteractionLayer interactionLayer;
-		WorldView cubeWorld;
-
-		public TumbleCubeControl(InteractionLayer interactionLayer)
-			: base(100, 100)
-		{
-			this.interactionLayer = interactionLayer;
-
-			TextureFace(cube.Faces[0], "Top");
-			TextureFace(cube.Faces[1], "Left", Matrix4X4.CreateRotationZ(MathHelper.Tau / 4));
-			TextureFace(cube.Faces[2], "Right", Matrix4X4.CreateRotationZ(-MathHelper.Tau/4));
-			TextureFace(cube.Faces[3], "Bottom", Matrix4X4.CreateRotationZ(MathHelper.Tau / 2));
-			TextureFace(cube.Faces[4], "Back", Matrix4X4.CreateRotationZ(MathHelper.Tau / 2));
-			TextureFace(cube.Faces[5], "Front");
-
-			cubeTraceData = cube.CreateTraceData();
-		}
-
-		public override void OnDraw(Graphics2D graphics2D)
-		{
-			var screenSpcaeBounds = this.TransformToScreenSpace(LocalBounds);
-			cubeWorld = new WorldView(screenSpcaeBounds.Width, screenSpcaeBounds.Height);
-
-			var forward = -Vector3.UnitZ;
-			var directionForward = Vector3.TransformNormal(forward, interactionLayer.World.InverseModelviewMatrix);
-
-			var up = Vector3.UnitY;
-			var directionUp = Vector3.TransformNormal(up, interactionLayer.World.InverseModelviewMatrix);
-			cubeWorld.RotationMatrix = Matrix4X4.LookAt(Vector3.Zero, directionForward, directionUp);
-
-			InteractionLayer.SetGlContext(cubeWorld, screenSpcaeBounds, lighting);
-			GLHelper.Render(cube, Color.White, Matrix4X4.Identity, RenderTypes.Shaded);
-			InteractionLayer.UnsetGlContext();
-
-			base.OnDraw(graphics2D);
-		}
-
-		public override void OnMouseDown(MouseEventArgs mouseEvent)
-		{
-			base.OnMouseDown(mouseEvent);
-
-			Ray ray = cubeWorld.GetRayForLocalBounds(mouseEvent.Position);
-			IntersectInfo info = cubeTraceData.GetClosestIntersection(ray);
-
-			if (info != null)
-			{
-				var normal = ((TriangleShape)info.closestHitObject).Plane.PlaneNormal;
-				var directionForward = -new Vector3(normal);
-
-				var directionUp = Vector3.UnitY;
-				if (directionForward.Equals(Vector3.UnitX, .001))
-				{
-					directionUp = Vector3.UnitZ;
-				}
-				else if (directionForward.Equals(-Vector3.UnitX, .001))
-				{
-					directionUp = Vector3.UnitZ;
-				}
-				else if (directionForward.Equals(Vector3.UnitY, .001))
-				{
-					directionUp = Vector3.UnitZ;
-				}
-				else if (directionForward.Equals(-Vector3.UnitY, .001))
-				{
-					directionUp = Vector3.UnitZ;
-				}
-				else if (directionForward.Equals(Vector3.UnitZ, .001))
-				{
-					directionUp = -Vector3.UnitY;
-				}
-
-				var look = Matrix4X4.LookAt(Vector3.Zero, directionForward, directionUp);
-
-				var start = new Quaternion(interactionLayer.World.RotationMatrix);
-				var end = new Quaternion(look);
-
-				Task.Run(() =>
-				{
-					double duration = .25;
-					var timer = Stopwatch.StartNew();
-					var time = timer.Elapsed.TotalSeconds;
-					while (time < duration)
-					{
-						var current = Quaternion.Slerp(start, end, time / duration);
-						UiThread.RunOnIdle(() =>
-						{
-							interactionLayer.World.RotationMatrix = Matrix4X4.CreateRotation(current);
-							Invalidate();
-						});
-						time = timer.Elapsed.TotalSeconds;
-						Thread.Sleep(10);
-					}
-					interactionLayer.World.RotationMatrix = Matrix4X4.CreateRotation(end);
-					Invalidate();
-				});
-			}
-		}
-
-		public override void OnMouseMove(MouseEventArgs mouseEvent)
-		{
-			// find the ray for this control
-			// check what face it hits
-			// mark that face to draw a highlight
-			base.OnMouseMove(mouseEvent);
-		}
-
-		public override void OnMouseUp(MouseEventArgs mouseEvent)
-		{
-			base.OnMouseUp(mouseEvent);
-
-			interactionLayer.Focus();
-		}
-
-		private static void TextureFace(Face face, string name, Matrix4X4? initialRotation = null)
-		{
-			ImageBuffer textureToUse = new ImageBuffer(256, 256);
-			var frontGraphics = textureToUse.NewGraphics2D();
-			frontGraphics.Clear(Color.White);
-			frontGraphics.DrawString(name,
-				textureToUse.Width / 2,
-				textureToUse.Height / 2,
-				60,
-				justification: Agg.Font.Justification.Center,
-				baseline: Agg.Font.Baseline.BoundsCenter);
-			MeshHelper.PlaceTextureOnFace(face, textureToUse, MeshHelper.GetMaxFaceProjection(face, textureToUse, initialRotation));
-		}
-	}
-
 	public class InteractionLayer : GuiWidget, IInteractionVolumeContext
 	{
 		private int volumeIndexWithMouseDown = -1;
@@ -211,8 +60,6 @@ namespace MatterHackers.MatterControl.PartPreviewWindow
 
 		private UndoBuffer undoBuffer;
 
-		private Action notifyPartChanged;
-
 		private LightingData lighting = new LightingData();
 
 		public InteractionLayer(WorldView world, UndoBuffer undoBuffer, InteractiveScene scene)
@@ -221,7 +68,6 @@ namespace MatterHackers.MatterControl.PartPreviewWindow
 			this.World = world;
 			this.InteractionVolumes = interactionVolumes;
 			this.undoBuffer = undoBuffer;
-			this.notifyPartChanged = notifyPartChanged;
 
 			var labelContainer = new GuiWidget();
 			labelContainer.Selectable = false;
@@ -406,7 +252,6 @@ namespace MatterHackers.MatterControl.PartPreviewWindow
 			if (this.Scene.HasSelection && this.Scene.SelectedItem.Matrix != originalTransform)
 			{
 				this.undoBuffer.Add(new TransformUndoCommand(Scene.SelectedItem, originalTransform, Scene.SelectedItem.Matrix));
-				this.notifyPartChanged?.Invoke();
 			}
 		}
